init folders and only use it if set.
2005-08-22  Not Zed  <NotZed@Ximian.com>

        * camel-store.c (camel_store_rename_folder): init folders and only
        use it if set.

        * camel-offline-journal.c (CAMEL_OFFLINE_JOURNAL_GET_CLASS):
        move redefined macro to .h.

        * camel-gpg-context.c (gpg_decrypt): cast args properly.

        * camel-text-index.c (camel_text_index_validate): cast to int for
        printf.

        * camel-object.c (co_metadata_get): assign metalast to remove
        spurious warning.

+++ camel/camel-text-index.c
@@ -1337,7 +1337,7 @@ camel_text_index_validate(CamelTextIndex *idx)
 				printf("Warning, read failed for word '%s', at data '%d'\n", word, data);
 				data = 0;
 			} else {
-				printf("(%d)\n", count);
+				printf("(%d)\n", (int)count);
 				g_free(records);
 			}
 		}
